Agus

+++ GccApplication3/Include/Header/DHT11.h
@@ -4,6 +4,6 @@
 #include <avr/io.h>
 
 void DHT11_Init(void);
-uint8_t DHT11_Read(int *temperature_1, int *temperature_2, int *humidity_1, int *humidity_2);
+uint8_t DHT11_Read(uint8_t *temperature_1, uint8_t *temperature_2, uint8_t *humidity_1, uint8_t *humidity_2);
 
 #endif /* DHT11_H_ */
